requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
PyMySQL==1.1.0
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.21
cryptography==41.0.7 